feat/create user on db and make verifications

+++ src/app/(dashboard)/_actions/categories.ts
@@ -23,31 +23,42 @@ export async function CreateCategory(form: CreateCategoriesSchemaType) {
     redirect('/sign-in')
   }
 
-  const { name, icon, type } = parsedBody.data
-  
-  const existCategoryName = await prisma.category.findMany({
+  // VERIFICAR SE O USUARIO EXISTE NO BD
+  const existUserOnDb = await prisma.user.findFirst({
     where: {
-      name,
-      userId: user.id
+      clerkUserId: user.id,
     },
   })
-  if (existCategoryName.length > 0) {
-    throw new Error('Already exists category with same name...')
-  }
 
-  return await prisma.category.create({
-    data: {
-      userId: user.id,
-      name,
-      icon,
-      type,
-    } as {
-      userId: string;
-      name: string;
-      icon: string;
-      type: "income" | "expanse";
+  const { name, icon, type } = parsedBody.data
+
+  if (existUserOnDb) {
+    const existCategoryName = await prisma.category.findMany({
+      where: {
+        name,
+        userId: existUserOnDb.id,
+        type
+      },
+    })
+  
+    if (existCategoryName.length > 0) {
+      throw new Error('Already exists category with same name...')
     }
-  })
+    
+    return await prisma.category.create({
+      data: {
+        userId: existUserOnDb.id,
+        name,
+        icon,
+        type,
+      } as {
+        userId: string;
+        name: string;
+        icon: string;
+        type: "income" | "expanse";
+      }
+    })
+  }
 }
 
 export async function DeleteCategory(form: DeleteCategorySchemaType) {
@@ -60,13 +71,24 @@ export async function DeleteCategory(form: DeleteCategorySchemaType) {
     redirect('/sign-in')
   }
 
-  return await prisma.category.delete({
+  // VERIFICAR SE O USUARIO EXISTE NO BD
+  const existUserOnDb = await prisma.user.findFirst({
     where: {
-      name_userId_type: {
-        userId: user.id,
-        name: parsedBody.data.name,
-        type: parsedBody.data.type,
-      },
+      clerkUserId: user.id,
     },
   })
+
+  if(existUserOnDb) {
+    return await prisma.category.delete({
+      where: {
+        name_userId_type: {
+          userId: existUserOnDb.id,
+          name: parsedBody.data.name,
+          type: parsedBody.data.type,
+        },
+      },
+    })
+  } else {
+    redirect('/sign-in')
+  }
 }
